backend/requirements.txt
# Django core
Django==5.1.4
wagtail==6.3.3
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
#psycopg2-binary

# Environment variables
python-decouple==3.8

# Images
Pillow==11.0.0

# Static files (production)
whitenoise==6.8.2

# WSGI server (production)
gunicorn==23.0.0

# Wagtail extras
wagtail-headless-preview==0.8.0

# Timezone support (used in RSS feed view)
pytz==2024.2

# Utilities
django-extensions==3.2.3
